+++ Packs/CortexXDR/Integrations/PaloAltoNetworks_XDR/CHANGELOG.md
@@ -1,4 +1,12 @@
 ## [Unreleased]
+- Added 6 commands.
+  - ***xdr-blacklist-files***
+  - ***xdr-whitelist-files***
+  - ***xdr-quarantine-files***
+  - ***xdr-get-quarantine-status***
+  - ***xdr-restore-file***
+  - ***xdr-endpoint-scan***
+- Added get-quarantine-file-status playbook.
 - Fixed a bug in the ***xdr-get-endpoint*** command where only the last endpoint was displayed in context.
 
 ## [20.4.1] - 2020-04-29
@@ -6,7 +14,7 @@
 
 ## [20.4.0] - 2020-04-14
   - Fixed the issue where the ***xdr-isolate-endpoint*** command was failing when:  
-    - The endpoint was disconnected.  
+    - The endpoint was disconnected. 
     - The isolation was still pending.  
     - The isolation cancellation was still pending.
   - Fixed the issue where ***xdr-unisolate-endpoint*** was failing when: